Create dedicated hook for being fragmentation aware
The styling needs to be constantly aware of what the octave groups are,
even when fragmentation is turned off. Solving this with parameters to a
single function was not as clear as it is now.
