+++ packages/harpguru-core/src/hooks/index.ts
@@ -7,3 +7,4 @@ export { useOptionSizes } from './use-option-sizes'
 export { useSizes } from './use-sizes'
 export { useIsZoomedColumnBounds } from './use-is-zoomed-columnbounds'
 export { useOctaveColumnGroups } from './use-octave-column-groups'
+export { useOctaveColumnGroupsFragmentAware } from './use-octave-column-groups-fragment-aware'
