+++ packages/harpguru-core/src/hooks/use-octave-column-groups/index.ts
@@ -1,1 +1,2 @@
 export { useOctaveColumnGroups } from './use-octave-column-groups'
+export type { ColumnRanges } from './get-octave-column-groups'
