+++ packages/harpguru-core/src/hooks/use-octave-column-groups/use-octave-column-groups.test.ts
@@ -9,15 +9,8 @@ const mockUseGlobal = useGlobal as jest.Mock
 
 test('Identifies the columns grouped by octave', () => {
   mockUseGlobal.mockImplementation(buildMockUseGlobalImplementation({}))
-  const octaveColumnGroups = useOctaveColumnGroups('harpface1', true)
+  const octaveColumnGroups = useOctaveColumnGroups('harpface1')
   const expectedOctaveGroups = [[0], [1, 2, 3, 4], [5, 6, 7], [8, 9]]
-
-  expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
-})
-
-test('Returns an ungrouped set if the harp fragmentation is off', () => {
-  const octaveColumnGroups = useOctaveColumnGroups('harpface1', false)
-  const expectedOctaveGroups = [[0, 1, 2, 3, 4, 5, 6, 7, 8, 9]]
 
   expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
 })
@@ -28,23 +21,11 @@ test('Returns a bounded grouped set', () => {
       sourceColumnBounds: [2, 7],
     })
   )
-  const octaveColumnGroups = useOctaveColumnGroups('harpface1', true)
+  const octaveColumnGroups = useOctaveColumnGroups('harpface1')
   const expectedOctaveGroups = [
     [2, 3, 4],
     [5, 6, 7],
   ]
-
-  expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
-})
-
-test('Returns a bounded ungrouped set', () => {
-  mockUseGlobal.mockImplementation(
-    buildMockUseGlobalImplementation({
-      sourceColumnBounds: [3, 8],
-    })
-  )
-  const octaveColumnGroups = useOctaveColumnGroups('harpface1', false)
-  const expectedOctaveGroups = [[3, 4, 5, 6, 7, 8]]
 
   expect(octaveColumnGroups).toStrictEqual(expectedOctaveGroups)
 })
